pyproject.toml
[project]
name = "neo4j-mpgnn-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

# Core dependencies that are common across all platforms.
# Torch and PyG related libraries are defined in platform-specific "extras" below.
dependencies = [
    "neo4j-rust-ext>=6.1.0.0",
    "torch_geometric==2.7.0",
    "scipy==1.13.1",
    "numpy==1.26.4",
    "pyarrow==22.0.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "dotenv>=0.9.9",
    "graphdatascience>=1.20",
    "notebook>=7.5.4",
    "ogb>=1.3.6",
    "tqdm>=4.67.3",
    "torch==2.7.1",
    "torch_cluster==1.6.3",
    "torch_sparse==0.6.18",
    "torch_scatter==2.1.2",
    "pyg-lib>=0.5.0",
]

[tool.uv.extra-build-dependencies]
pyg-lib = ["torch"]
torch-cluster = ["torch"]
torch-scatter = ["torch"]
torch-sparse = ["torch"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu"},
]
pyg-lib = [
    { index = "pytorch-geometric-cpu" },
]
torch_cluster = [
    { index = "pytorch-geometric-cpu" },
]
torch_scatter = [
    { index = "pytorch-geometric-cpu" },
]
torch_sparse = [
    { index = "pytorch-geometric-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

#[[tool.uv.index]]
#name = "pytorch-cu126"
#url = "https://download.pytorch.org/whl/cu126"
#explicit = true

[[tool.uv.index]]
name = "pytorch-geometric-cpu"
url = "https://data.pyg.org/whl/torch-2.7.1+cpu.html"
explicit = true
format = "flat"

#[[tool.uv.index]]
#name = "pytorch-geometric-cu126"
#url = "https://data.pyg.org/whl/torch-2.7.1+cu126.html"
#explicit = true
#format = "flat"


[tool.uv]
required-environments = [
  "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
